Authentication applied

+++ app/graphql/mutations/delete_book.rb
@@ -3,19 +3,17 @@ class Mutations::DeleteBook < Mutations::BaseMutation
     description "Delete Book"
 
     argument :id, ID, required: true
-  
-    field :book, Types::BookType, null: false
-    field :errors, [String], null: false
-
+    type Types::BookType
+    
     def resolve(id:)
-        book = Book.find(id)
+      unless context[:current_user].nil?
+        author = context[:current_user]
+        book = author.books.find(id)
         book.chapters.clear
         book.destroy!
-        if(book.errors.blank?)
-          {book: book, errors: []}
-        else
-          {book: [], errors: book.errors.full_messages}
-        end
+      else
+        raise GraphQL::ExecutionError, "Authentication failed, Sign in first"
+      end
     end
 
 end
